Optimize manual upload E2E tests by consolidating into single file
Merged manual-upload.spec.ts, manual-upload-success.spec.ts, and
manual-upload-summary.spec.ts into a single comprehensive test file.
Reduced test count from 60+ to 17 consolidated tests while maintaining
100% coverage through meaningful end-to-end flows.

Key improvements:
- Combined related tests (form display, validation, navigation)
- Reduced duplicate page loads and authentication calls
- Added helper functions for common test workflows
- Fixed flaky navigation test using simplified approach
- Improved test efficiency by reusing form data across validation tests

All 17 tests passing with improved execution time (18.8s vs previous ~1.1m).
