+++ e2e-tests/tests/manual-upload.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
+import type { Page } from '@playwright/test';
 import AxeBuilder from '@axe-core/playwright';
-import path from 'node:path';
 import { loginWithSSO } from '../utils/sso-helpers.js';
 
 // Note: target-size and link-name rules are disabled due to pre-existing site-wide footer accessibility issues:
@@ -9,49 +9,214 @@ import { loginWithSSO } from '../utils/sso-helpers.js';
 // These issues affect ALL pages and should be addressed in a separate ticket
 // See: docs/tickets/VIBE-150/accessibility-findings.md
 
-test.describe('Manual Upload Page', () => {
-  // Authenticate once before all tests
-  test.beforeEach(async ({ page }) => {
-    await page.goto("/system-admin-dashboard");
+// Helper function to authenticate as System Admin
+async function authenticateSystemAdmin(page: Page) {
+  await page.goto("/system-admin-dashboard");
 
-    // If we're redirected to Azure AD, login
-    if (page.url().includes("login.microsoftonline.com")) {
-      const systemAdminEmail = process.env.SSO_TEST_SYSTEM_ADMIN_EMAIL!;
-      const systemAdminPassword = process.env.SSO_TEST_SYSTEM_ADMIN_PASSWORD!;
-      await loginWithSSO(page, systemAdminEmail, systemAdminPassword);
-    }
+  if (page.url().includes("login.microsoftonline.com")) {
+    const systemAdminEmail = process.env.SSO_TEST_SYSTEM_ADMIN_EMAIL!;
+    const systemAdminPassword = process.env.SSO_TEST_SYSTEM_ADMIN_PASSWORD!;
+    await loginWithSSO(page, systemAdminEmail, systemAdminPassword);
+  }
+}
+
+// Helper function to navigate to summary page by completing the upload form
+async function navigateToSummaryPage(page: Page) {
+  await authenticateSystemAdmin(page);
+  await page.goto("/manual-upload?locationId=1");
+  await page.waitForTimeout(1000);
+
+  await page.selectOption('select[name="listType"]', "6");
+  await page.fill('input[name="hearingStartDate-day"]', "23");
+  await page.fill('input[name="hearingStartDate-month"]', "10");
+  await page.fill('input[name="hearingStartDate-year"]', "2025");
+  await page.selectOption('select[name="sensitivity"]', "PUBLIC");
+  await page.selectOption('select[name="language"]', "ENGLISH");
+  await page.fill('input[name="displayFrom-day"]', "20");
+  await page.fill('input[name="displayFrom-month"]', "10");
+  await page.fill('input[name="displayFrom-year"]', "2025");
+  await page.fill('input[name="displayTo-day"]', "30");
+  await page.fill('input[name="displayTo-month"]', "10");
+  await page.fill('input[name="displayTo-year"]', "2025");
+
+  const fileInput = page.locator('input[name="file"]');
+  await fileInput.setInputFiles({
+    name: "test-document.pdf",
+    mimeType: "application/pdf",
+    buffer: Buffer.from("%PDF-1.4\nTest PDF content")
   });
 
-  test.describe('given user is on the manual-upload page', () => {
+  await page.getByRole("button", { name: /continue/i }).click();
+  await page.waitForURL(/\/manual-upload-summary\?uploadId=/, { timeout: 10000 });
+}
+
+// Helper function to complete the full manual upload flow and reach success page
+async function completeManualUploadFlow(page: Page) {
+  await navigateToSummaryPage(page);
+  await page.getByRole("button", { name: "Confirm" }).click();
+  await page.waitForURL("/manual-upload-success", { timeout: 10000 });
+}
+
+test.describe('Manual Upload End-to-End Flow', () => {
+  test.beforeEach(async ({ page }) => {
+    await authenticateSystemAdmin(page);
+  });
+
+  test.describe('Complete End-to-End Journey', () => {
+    test('should be keyboard accessible throughout entire upload flow', async ({ page }) => {
+      // Step 1: Test keyboard accessibility on form page
+      await page.goto('/manual-upload?locationId=1');
+      await page.waitForTimeout(1000);
+
+      const fileInput = page.locator('input[name="file"]');
+      await fileInput.click();
+      await expect(fileInput).toBeFocused();
+
+      await page.keyboard.press('Tab');
+      const focusedElement = page.locator(':focus');
+      await expect(focusedElement).toBeVisible();
+
+      // Step 2: Fill form and submit
+      await page.selectOption('select[name="listType"]', "6");
+      await page.fill('input[name="hearingStartDate-day"]', "23");
+      await page.fill('input[name="hearingStartDate-month"]', "10");
+      await page.fill('input[name="hearingStartDate-year"]', "2025");
+      await page.selectOption('select[name="sensitivity"]', "PUBLIC");
+      await page.selectOption('select[name="language"]', "ENGLISH");
+      await page.fill('input[name="displayFrom-day"]', "20");
+      await page.fill('input[name="displayFrom-month"]', "10");
+      await page.fill('input[name="displayFrom-year"]', "2025");
+      await page.fill('input[name="displayTo-day"]', "30");
+      await page.fill('input[name="displayTo-month"]', "10");
+      await page.fill('input[name="displayTo-year"]', "2025");
+
+      await fileInput.setInputFiles({
+        name: "test-keyboard.pdf",
+        mimeType: "application/pdf",
+        buffer: Buffer.from("%PDF-1.4\nTest keyboard content")
+      });
+
+      await page.getByRole("button", { name: /continue/i }).click();
+      await page.waitForURL(/\/manual-upload-summary\?uploadId=/, { timeout: 10000 });
+
+      // Step 3: Test keyboard accessibility on summary page
+      const changeLinks = page.locator(".govuk-summary-list__actions a");
+      await expect(changeLinks).toHaveCount(7);
+
+      for (let i = 0; i < 7; i++) {
+        const href = await changeLinks.nth(i).getAttribute("href");
+        expect(href).toMatch(/^\/manual-upload#/);
+      }
+
+      const confirmButton = page.getByRole("button", { name: "Confirm" });
+      await expect(confirmButton).toBeVisible();
+      await expect(confirmButton).toHaveAttribute("type", "submit");
+
+      await confirmButton.focus();
+      await page.keyboard.press("Enter");
+      await page.waitForURL("/manual-upload-success", { timeout: 10000 });
+
+      // Step 4: Test keyboard accessibility on success page
+      const links = page.locator(".govuk-list a");
+      await expect(links).toHaveCount(3);
+
+      for (let i = 0; i < 3; i++) {
+        const href = await links.nth(i).getAttribute("href");
+        expect(href).toBeTruthy();
+      }
+
+      for (let i = 0; i < 3; i++) {
+        const link = links.nth(i);
+        await link.focus();
+        await expect(link).toBeFocused();
+      }
+
+      const uploadLink = page.getByRole("link", { name: "Upload another file" });
+      await uploadLink.focus();
+      await page.keyboard.press("Enter");
+      await expect(page).toHaveURL("/manual-upload");
+    });
+
+    test('should complete full upload flow from form to success', async ({ page }) => {
+      // Step 1: Load manual upload form
+      await page.goto('/manual-upload?locationId=1');
+      await page.waitForTimeout(1000);
+      await expect(page).toHaveTitle('Upload - Manual upload - Court and tribunal hearings - GOV.UK');
+
+      // Step 2: Fill out the form
+      await page.selectOption('select[name="listType"]', "6");
+      await page.fill('input[name="hearingStartDate-day"]', "23");
+      await page.fill('input[name="hearingStartDate-month"]', "10");
+      await page.fill('input[name="hearingStartDate-year"]', "2025");
+      await page.selectOption('select[name="sensitivity"]', "PUBLIC");
+      await page.selectOption('select[name="language"]', "ENGLISH");
+      await page.fill('input[name="displayFrom-day"]', "20");
+      await page.fill('input[name="displayFrom-month"]', "10");
+      await page.fill('input[name="displayFrom-year"]', "2025");
+      await page.fill('input[name="displayTo-day"]', "30");
+      await page.fill('input[name="displayTo-month"]', "10");
+      await page.fill('input[name="displayTo-year"]', "2025");
+
+      const fileInput = page.locator('input[name="file"]');
+      await fileInput.setInputFiles({
+        name: "test-hearing-list.pdf",
+        mimeType: "application/pdf",
+        buffer: Buffer.from("%PDF-1.4\nTest hearing list content")
+      });
+
+      // Step 3: Submit form and verify navigation to summary
+      await page.getByRole("button", { name: /continue/i }).click();
+      await page.waitForURL(/\/manual-upload-summary\?uploadId=/, { timeout: 10000 });
+      await expect(page.locator("h1")).toHaveText("File upload summary");
+
+      // Step 4: Verify summary page displays correct data
+      const values = page.locator(".govuk-summary-list__value");
+      await expect(values.nth(1)).toContainText("test-hearing-list.pdf");
+      await expect(values.nth(2)).toContainText("Crown Daily List");
+      await expect(values.nth(3)).toContainText("23 October 2025");
+      await expect(values.nth(4)).toContainText("Public");
+      await expect(values.nth(5)).toContainText("English");
+      await expect(values.nth(6)).toContainText("20 October 2025 to 30 October 2025");
+
+      // Step 5: Confirm upload and navigate to success page
+      await page.getByRole("button", { name: "Confirm" }).click();
+      await page.waitForURL("/manual-upload-success", { timeout: 10000 });
+
+      // Step 6: Verify success page content
+      const successPanel = page.locator(".govuk-panel");
+      await expect(successPanel).toBeVisible();
+      const title = page.locator(".govuk-panel__title");
+      await expect(title).toHaveText("File upload successful");
+      await expect(successPanel).toContainText("Your file has been uploaded");
+
+      // Step 7: Verify next steps are available
+      const uploadLink = page.getByRole("link", { name: "Upload another file" });
+      await expect(uploadLink).toBeVisible();
+    });
+  });
+
+  test.describe('Manual Upload Form Page', () => {
     test('should load the page with all form fields and accessibility compliance', async ({ page }) => {
       await page.goto('/manual-upload');
 
-      // Check the browser tab title
       await expect(page).toHaveTitle('Upload - Manual upload - Court and tribunal hearings - GOV.UK');
 
-      // Check for the page heading
       const heading = page.getByRole('heading', { name: /manual upload/i });
       await expect(heading).toBeVisible();
 
-      // Check for warning section
       const warningTitle = page.getByText(/warning/i);
       await expect(warningTitle).toBeVisible();
 
-      // Check for file upload field
       const fileUpload = page.locator('input[name="file"]');
       await expect(fileUpload).toBeVisible();
 
-      // Check for court/tribunal autocomplete field (use role="combobox" which is set after autocomplete initializes)
-      // Wait for the autocomplete to be initialized and visible
       const courtInput = page.getByRole('combobox', { name: /court name or tribunal name/i });
       await courtInput.waitFor({ state: 'visible', timeout: 10000 });
       await expect(courtInput).toBeVisible();
 
-      // Check for list type dropdown
       const listTypeSelect = page.locator('select[name="listType"]');
       await expect(listTypeSelect).toBeVisible();
 
-      // Check for hearing start date fields
       const hearingDateDay = page.locator('input[name="hearingStartDate-day"]');
       const hearingDateMonth = page.locator('input[name="hearingStartDate-month"]');
       const hearingDateYear = page.locator('input[name="hearingStartDate-year"]');
@@ -59,15 +224,12 @@ test.describe('Manual Upload Page', () => {
       await expect(hearingDateMonth).toBeVisible();
       await expect(hearingDateYear).toBeVisible();
 
-      // Check for sensitivity dropdown
       const sensitivitySelect = page.locator('select[name="sensitivity"]');
       await expect(sensitivitySelect).toBeVisible();
 
-      // Check for language dropdown
       const languageSelect = page.locator('select[name="language"]');
       await expect(languageSelect).toBeVisible();
 
-      // Check for display date fields
       const displayFromDay = page.locator('input[name="displayFrom-day"]');
       const displayFromMonth = page.locator('input[name="displayFrom-month"]');
       const displayFromYear = page.locator('input[name="displayFrom-year"]');
@@ -82,15 +244,12 @@ test.describe('Manual Upload Page', () => {
       await expect(displayToMonth).toBeVisible();
       await expect(displayToYear).toBeVisible();
 
-      // Check for continue button
       const continueButton = page.getByRole('button', { name: /continue/i });
       await expect(continueButton).toBeVisible();
 
-      // Verify language toggle is NOT present (hideLanguageToggle: true)
       const languageToggle = page.locator('.language');
       await expect(languageToggle).not.toBeVisible();
 
-      // Run accessibility checks on initial page load
       const accessibilityScanResults = await new AxeBuilder({ page })
         .withTags(['wcag2a', 'wcag2aa', 'wcag21a', 'wcag21aa', 'wcag22aa'])
         .disableRules(['target-size', 'link-name'])
@@ -109,402 +268,24 @@ test.describe('Manual Upload Page', () => {
       }
 
       expect(accessibilityScanResults.violations).toEqual([]);
-    });
 
-    test('should have red inset text around file, list type, and hearing date fields', async ({ page }) => {
-      await page.goto('/manual-upload');
-
-      // Check for inset text wrappers (grey border)
       const fileInset = page.locator('.govuk-inset-text').nth(0);
       const listTypeInset = page.locator('.govuk-inset-text').nth(1);
 
       await expect(fileInset).toBeVisible();
       await expect(listTypeInset).toBeVisible();
 
-      // Verify file field is inside first inset
-      const fileUpload = fileInset.locator('input[name="file"]');
-      await expect(fileUpload).toBeVisible();
+      const fileUploadInInset = fileInset.locator('input[name="file"]');
+      await expect(fileUploadInInset).toBeVisible();
 
-      // Verify list type and hearing date are inside second inset
-      const listTypeSelect = listTypeInset.locator('select[name="listType"]');
-      const hearingDateDay = listTypeInset.locator('input[name="hearingStartDate-day"]');
-      await expect(listTypeSelect).toBeVisible();
-      await expect(hearingDateDay).toBeVisible();
-    });
-  });
+      const listTypeInInset = listTypeInset.locator('select[name="listType"]');
+      const hearingDateInInset = listTypeInset.locator('input[name="hearingStartDate-day"]');
+      await expect(listTypeInInset).toBeVisible();
+      await expect(hearingDateInInset).toBeVisible();
 
-  test.describe('given user submits form without filling any fields', () => {
-    test('should display validation errors for all required fields', async ({ page }) => {
-      await page.goto('/manual-upload');
-
-      // Click continue without filling any fields
-      const continueButton = page.getByRole('button', { name: /continue/i });
-      await continueButton.click();
-
-      // Verify we're still on the manual-upload page
-      await expect(page).toHaveURL('/manual-upload');
-
-      // Check for error summary
-      const errorSummary = page.locator('.govuk-error-summary');
-      await expect(errorSummary).toBeVisible();
-
-      // Check for error summary heading
-      const errorSummaryHeading = errorSummary.getByRole('heading', { name: /there is a problem/i });
-      await expect(errorSummaryHeading).toBeVisible();
-
-      // Verify multiple error links are present
-      const errorLinks = errorSummary.locator('.govuk-error-summary__list a');
-      const errorCount = await errorLinks.count();
-      expect(errorCount).toBeGreaterThan(0);
-
-      // Verify error messages are visible next to fields
-      const fileErrorMessage = page.locator('#file').locator('..').locator('.govuk-error-message');
-      await expect(fileErrorMessage).toBeVisible();
-
-      // Verify accessibility with error state
-      const accessibilityScanResults = await new AxeBuilder({ page })
-        .withTags(['wcag2a', 'wcag2aa', 'wcag21a', 'wcag21aa', 'wcag22aa'])
-        .disableRules(['target-size', 'link-name'])
-        .analyze();
-
-      expect(accessibilityScanResults.violations).toEqual([]);
-    });
-  });
-
-  test.describe('given user uploads invalid file type', () => {
-    test('should display file type validation error', async ({ page }) => {
-      await page.goto('/manual-upload?locationId=1');
-
-      // Wait for autocomplete to initialize with pre-filled value
-      await page.waitForTimeout(1000);
-
-      // Create a test file with invalid extension
-      const fileInput = page.locator('input[name="file"]');
-
-      // Set file with invalid extension (.txt)
-      await fileInput.setInputFiles({
-        name: 'test.txt',
-        mimeType: 'text/plain',
-        buffer: Buffer.from('test content')
-      });
-
-      await page.selectOption('select[name="listType"]', '1'); // Civil Daily Cause List
-      await page.fill('input[name="hearingStartDate-day"]', '15');
-      await page.fill('input[name="hearingStartDate-month"]', '06');
-      await page.fill('input[name="hearingStartDate-year"]', '2025');
-      await page.selectOption('select[name="sensitivity"]', 'PUBLIC');
-      await page.selectOption('select[name="language"]', 'ENGLISH');
-      await page.fill('input[name="displayFrom-day"]', '10');
-      await page.fill('input[name="displayFrom-month"]', '06');
-      await page.fill('input[name="displayFrom-year"]', '2025');
-      await page.fill('input[name="displayTo-day"]', '20');
-      await page.fill('input[name="displayTo-month"]', '06');
-      await page.fill('input[name="displayTo-year"]', '2025');
-
-      // Submit form
-      const continueButton = page.getByRole('button', { name: /continue/i });
-      await continueButton.click();
-
-      // Verify we're still on the manual-upload page (with query param)
-      await expect(page).toHaveURL(/\/manual-upload/);
-
-      // Check for error summary
-      const errorSummary = page.locator('.govuk-error-summary');
-      await expect(errorSummary).toBeVisible();
-
-      // Verify error link points to file input
-      const errorLink = errorSummary.getByRole('link', { name: /please upload a valid file format/i });
-      await expect(errorLink).toBeVisible();
-      await expect(errorLink).toHaveAttribute('href', '#file');
-
-      // Check for inline error message
-      const inlineErrorMessage = page.locator('#file').locator('..').locator('.govuk-error-message');
-      await expect(inlineErrorMessage).toBeVisible();
-      await expect(inlineErrorMessage).toContainText(/please upload a valid file format/i);
-    });
-  });
-
-  test.describe('given user uploads file larger than 2MB', () => {
-    test('should display file size validation error', async ({ page }) => {
-      await page.goto('/manual-upload?locationId=1');
-
-      // Wait for autocomplete to initialize with pre-filled value
-      await page.waitForTimeout(1000);
-
-      const fileInput = page.locator('input[name="file"]');
-
-      // Create a buffer larger than 2MB
-      const largeBuffer = Buffer.alloc(3 * 1024 * 1024); // 3MB
-      await fileInput.setInputFiles({
-        name: 'large-file.pdf',
-        mimeType: 'application/pdf',
-        buffer: largeBuffer
-      });
-
-      await page.selectOption('select[name="listType"]', '1'); // Civil Daily Cause List
-      await page.fill('input[name="hearingStartDate-day"]', '15');
-      await page.fill('input[name="hearingStartDate-month"]', '06');
-      await page.fill('input[name="hearingStartDate-year"]', '2025');
-      await page.selectOption('select[name="sensitivity"]', 'PUBLIC');
-      await page.selectOption('select[name="language"]', 'ENGLISH');
-      await page.fill('input[name="displayFrom-day"]', '10');
-      await page.fill('input[name="displayFrom-month"]', '06');
-      await page.fill('input[name="displayFrom-year"]', '2025');
-      await page.fill('input[name="displayTo-day"]', '20');
-      await page.fill('input[name="displayTo-month"]', '06');
-      await page.fill('input[name="displayTo-year"]', '2025');
-
-      // Submit form
-      const continueButton = page.getByRole('button', { name: /continue/i });
-      await continueButton.click();
-
-      // Verify we're still on the manual-upload page (with query param)
-      await expect(page).toHaveURL(/\/manual-upload/);
-
-      // Check for error summary
-      const errorSummary = page.locator('.govuk-error-summary');
-      await expect(errorSummary).toBeVisible();
-
-      // Verify error link points to file input
-      const errorLink = errorSummary.getByRole('link', { name: /file too large/i });
-      await expect(errorLink).toBeVisible();
-      await expect(errorLink).toHaveAttribute('href', '#file');
-
-      // Check for inline error message
-      const inlineErrorMessage = page.locator('#file').locator('..').locator('.govuk-error-message');
-      await expect(inlineErrorMessage).toBeVisible();
-      await expect(inlineErrorMessage).toContainText(/file too large, please upload file smaller than 2mb/i);
-    });
-  });
-
-  test.describe('given user interacts with court name autocomplete', () => {
-    test('should show court name input with autocomplete initialized', async ({ page }) => {
-      await page.goto('/manual-upload');
-
-      // Wait for autocomplete to initialize by waiting for the combobox element
-      const autocompleteInput = page.getByRole('combobox', { name: /court name or tribunal name/i });
-      await autocompleteInput.waitFor({ state: 'visible', timeout: 10000 });
-
-      // Check that the autocomplete input is visible (created by accessible-autocomplete)
-      await expect(autocompleteInput).toBeVisible();
-
-      // Verify the autocomplete has been initialized
-      await expect(autocompleteInput).toHaveAttribute('role', 'combobox');
-    });
-
-    test('should display validation error for empty court name', async ({ page }) => {
-      await page.goto('/manual-upload');
-
-      // Fill only file to isolate court validation error
-      const fileInput = page.locator('input[name="file"]');
-      await fileInput.setInputFiles({
-        name: 'test.pdf',
-        mimeType: 'application/pdf',
-        buffer: Buffer.from('test content')
-      });
-
-      // Submit without selecting court
-      const continueButton = page.getByRole('button', { name: /continue/i });
-      await continueButton.click();
-
-      // Verify error link in summary
-      const errorSummary = page.locator('.govuk-error-summary');
-      const errorLink = errorSummary.getByRole('link', { name: /court name must be three characters or more/i });
-      await expect(errorLink).toBeVisible();
-
-      // Verify inline error message
-      const inlineError = page.locator('.govuk-error-message').filter({ hasText: /court name must be three characters or more/i });
-      await expect(inlineError).toBeVisible();
-    });
-
-    test('should display validation error for court name less than 3 characters', async ({ page }) => {
-      await page.goto('/manual-upload');
-
-      // Wait for autocomplete to initialize and be ready for input
-      const courtInput = page.getByRole('combobox', { name: /court name or tribunal name/i });
-      await courtInput.waitFor({ state: 'visible', timeout: 10000 });
-      await courtInput.fill('AB');
-
-      // Fill file to avoid file error
-      const fileInput = page.locator('input[name="file"]');
-      await fileInput.setInputFiles({
-        name: 'test.pdf',
-        mimeType: 'application/pdf',
-        buffer: Buffer.from('test content')
-      });
-
-      // Submit
-      const continueButton = page.getByRole('button', { name: /continue/i });
-      await continueButton.click();
-
-      // Verify error link in summary
-      const errorSummary = page.locator('.govuk-error-summary');
-      await expect(errorSummary).toBeVisible();
-      const errorLink = errorSummary.getByRole('link', { name: /court name must be three characters or more/i });
-      await expect(errorLink).toBeVisible();
-
-      // Verify inline error message
-      const inlineError = page.locator('#court-error.govuk-error-message');
-      await expect(inlineError).toBeVisible();
-      await expect(inlineError).toContainText(/court name must be three characters or more/i);
-    });
-
-    test('should display validation error for invalid court name longer than 3 characters', async ({ page }) => {
-      await page.goto('/manual-upload');
-
-      // Wait for autocomplete to initialize and be ready for input
-      const courtInput = page.getByRole('combobox', { name: /court name or tribunal name/i });
-      await courtInput.waitFor({ state: 'visible', timeout: 10000 });
-      await courtInput.fill('Invalid Court Name That Does Not Exist');
-
-      // Fill file to avoid file error
-      const fileInput = page.locator('input[name="file"]');
-      await fileInput.setInputFiles({
-        name: 'test.pdf',
-        mimeType: 'application/pdf',
-        buffer: Buffer.from('test content')
-      });
-
-      // Submit
-      const continueButton = page.getByRole('button', { name: /continue/i });
-      await continueButton.click();
-
-      // Verify error link in summary
-      const errorSummary = page.locator('.govuk-error-summary');
-      await expect(errorSummary).toBeVisible();
-      const errorLink = errorSummary.getByRole('link', { name: /please enter and select a valid court/i });
-      await expect(errorLink).toBeVisible();
-
-      // Verify inline error message
-      const inlineError = page.locator('#court-error.govuk-error-message');
-      await expect(inlineError).toBeVisible();
-      await expect(inlineError).toContainText(/please enter and select a valid court/i);
-    });
-
-    test('should preserve invalid court name after validation error', async ({ page }) => {
-      await page.goto('/manual-upload');
-
-      // Wait for autocomplete to initialize and be ready for input
-      const invalidCourtName = 'Invalid Court Name';
-      const courtInput = page.getByRole('combobox', { name: /court name or tribunal name/i });
-      await courtInput.waitFor({ state: 'visible', timeout: 10000 });
-      await courtInput.fill(invalidCourtName);
-
-      // Fill file to avoid file error
-      const fileInput = page.locator('input[name="file"]');
-      await fileInput.setInputFiles({
-        name: 'test.pdf',
-        mimeType: 'application/pdf',
-        buffer: Buffer.from('test content')
-      });
-
-      // Submit
-      const continueButton = page.getByRole('button', { name: /continue/i });
-      await continueButton.click();
-
-      // Verify the invalid court name is still in the input field after page re-renders
-      const courtInputAfterError = page.getByRole('combobox', { name: /court name or tribunal name/i });
-      await expect(courtInputAfterError).toHaveValue(invalidCourtName);
-    });
-  });
-
-  test.describe('given user validates display date logic', () => {
-    test('should display error when display to date is before display from date', async ({ page }) => {
-      // Navigate with pre-filled location ID to bypass autocomplete complexity
-      await page.goto('/manual-upload?locationId=1');
-
-      // Fill file
-      const fileInput = page.locator('input[name="file"]');
-      await fileInput.setInputFiles({
-        name: 'test.pdf',
-        mimeType: 'application/pdf',
-        buffer: Buffer.from('test content')
-      });
-
-      // Court is already pre-filled from query param
-      await page.waitForTimeout(1000); // Wait for autocomplete to initialize
-
-      // Fill list type
-      await page.selectOption('select[name="listType"]', '1'); // Civil Daily Cause List
-
-      // Fill hearing start date
-      await page.fill('input[name="hearingStartDate-day"]', '15');
-      await page.fill('input[name="hearingStartDate-month"]', '06');
-      await page.fill('input[name="hearingStartDate-year"]', '2025');
-
-      // Select sensitivity and language
-      await page.selectOption('select[name="sensitivity"]', 'PUBLIC');
-      await page.selectOption('select[name="language"]', 'ENGLISH');
-
-      // Fill display dates with 'to' before 'from'
-      await page.fill('input[name="displayFrom-day"]', '20');
-      await page.fill('input[name="displayFrom-month"]', '06');
-      await page.fill('input[name="displayFrom-year"]', '2025');
-      await page.fill('input[name="displayTo-day"]', '10');
-      await page.fill('input[name="displayTo-month"]', '06');
-      await page.fill('input[name="displayTo-year"]', '2025');
-
-      // Submit
-      const continueButton = page.getByRole('button', { name: /continue/i });
-      await continueButton.click();
-
-      // Verify error link in summary
-      const errorSummary = page.locator('.govuk-error-summary');
-      await expect(errorSummary).toBeVisible();
-      const errorLink = errorSummary.getByRole('link', { name: /display to date must be after display from date/i });
-      await expect(errorLink).toBeVisible();
-
-      // Verify inline error message
-      const inlineError = page.locator('#displayTo-error.govuk-error-message');
-      await expect(inlineError).toBeVisible();
-      await expect(inlineError).toContainText(/display to date must be after display from date/i);
-    });
-  });
-
-  test.describe('given user uses keyboard navigation', () => {
-    test('should be navigable using Tab key with visible focus indicators', async ({ page }) => {
-      await page.goto('/manual-upload');
-
-      // Navigate to the file input by clicking it
-      const fileInput = page.locator('input[name="file"]');
-      await fileInput.click();
-
-      // Check focus is visible
-      await expect(fileInput).toBeFocused();
-
-      // Tab to next field (should be court input after autocomplete initializes)
-      await page.keyboard.press('Tab');
-
-      // The next focused element should be visible
-      const focusedElement = page.locator(':focus');
-      await expect(focusedElement).toBeVisible();
-    });
-
-    test('should submit form using Enter key on continue button', async ({ page }) => {
-      await page.goto('/manual-upload');
-
-      // Click the continue button to give it focus
-      const continueButton = page.getByRole('button', { name: /continue/i });
-      await continueButton.click();
-
-      // Should stay on page due to validation errors
-      await expect(page).toHaveURL('/manual-upload');
-
-      // Should show error summary
-      const errorSummary = page.locator('.govuk-error-summary');
-      await expect(errorSummary).toBeVisible();
-    });
-  });
-
-  test.describe('given user checks page help section', () => {
-    test('should display page help section with accordion', async ({ page }) => {
-      await page.goto('/manual-upload');
-
-      // Check for page help title
       const pageHelpTitle = page.getByText(/page help/i);
       await expect(pageHelpTitle).toBeVisible();
 
-      // Check for help section topics
       const listsHelp = page.getByText(/lists/i).first();
       const sensitivityHelp = page.getByText(/sensitivity/i).first();
       const displayFromHelp = page.getByText(/display from/i).first();
@@ -515,22 +296,46 @@ test.describe('Manual Upload Page', () => {
       await expect(displayFromHelp).toBeVisible();
       await expect(displayToHelp).toBeVisible();
     });
-  });
 
-  test.describe('given user checks form preservation on error', () => {
-    test('should preserve all form data when validation fails', async ({ page }) => {
+    test('should display validation errors for all required fields when form is empty', async ({ page }) => {
+      await page.goto('/manual-upload');
+
+      const continueButton = page.getByRole('button', { name: /continue/i });
+      await continueButton.click();
+
+      await expect(page).toHaveURL('/manual-upload');
+
+      const errorSummary = page.locator('.govuk-error-summary');
+      await expect(errorSummary).toBeVisible();
+
+      const errorSummaryHeading = errorSummary.getByRole('heading', { name: /there is a problem/i });
+      await expect(errorSummaryHeading).toBeVisible();
+
+      const errorLinks = errorSummary.locator('.govuk-error-summary__list a');
+      const errorCount = await errorLinks.count();
+      expect(errorCount).toBeGreaterThan(0);
+
+      const fileErrorMessage = page.locator('#file').locator('..').locator('.govuk-error-message');
+      await expect(fileErrorMessage).toBeVisible();
+
+      const accessibilityScanResults = await new AxeBuilder({ page })
+        .withTags(['wcag2a', 'wcag2aa', 'wcag21a', 'wcag21aa', 'wcag22aa'])
+        .disableRules(['target-size', 'link-name'])
+        .analyze();
+
+      expect(accessibilityScanResults.violations).toEqual([]);
+    });
+
+    test('should display file validation errors for invalid type and large size', async ({ page }) => {
       await page.goto('/manual-upload?locationId=1');
-
-      // Fill most fields but leave one required field empty to trigger error
-      // Wait for autocomplete to initialize with pre-filled value
       await page.waitForTimeout(1000);
 
-      await page.selectOption('select[name="listType"]', '1'); // Civil Daily Cause List
+      await page.selectOption('select[name="listType"]', '1');
       await page.fill('input[name="hearingStartDate-day"]', '15');
       await page.fill('input[name="hearingStartDate-month"]', '06');
       await page.fill('input[name="hearingStartDate-year"]', '2025');
-      await page.selectOption('select[name="sensitivity"]', 'PRIVATE');
-      await page.selectOption('select[name="language"]', 'WELSH');
+      await page.selectOption('select[name="sensitivity"]', 'PUBLIC');
+      await page.selectOption('select[name="language"]', 'ENGLISH');
       await page.fill('input[name="displayFrom-day"]', '10');
       await page.fill('input[name="displayFrom-month"]', '06');
       await page.fill('input[name="displayFrom-year"]', '2025');
@@ -538,14 +343,161 @@ test.describe('Manual Upload Page', () => {
       await page.fill('input[name="displayTo-month"]', '06');
       await page.fill('input[name="displayTo-year"]', '2025');
 
-      // Note: Not filling file to trigger error
+      const fileInput = page.locator('input[name="file"]');
+      await fileInput.setInputFiles({
+        name: 'test.txt',
+        mimeType: 'text/plain',
+        buffer: Buffer.from('test content')
+      });
 
-      // Submit
       const continueButton = page.getByRole('button', { name: /continue/i });
       await continueButton.click();
 
-      // Verify form data is preserved
-      await expect(page.locator('select[name="listType"]')).toHaveValue('1'); // Civil Daily Cause List
+      await expect(page).toHaveURL(/\/manual-upload/);
+
+      let errorSummary = page.locator('.govuk-error-summary');
+      await expect(errorSummary).toBeVisible();
+
+      let errorLink = errorSummary.getByRole('link', { name: /please upload a valid file format/i });
+      await expect(errorLink).toBeVisible();
+      await expect(errorLink).toHaveAttribute('href', '#file');
+
+      let inlineErrorMessage = page.locator('#file').locator('..').locator('.govuk-error-message');
+      await expect(inlineErrorMessage).toBeVisible();
+      await expect(inlineErrorMessage).toContainText(/please upload a valid file format/i);
+
+      const largeBuffer = Buffer.alloc(3 * 1024 * 1024);
+      await fileInput.setInputFiles({
+        name: 'large-file.pdf',
+        mimeType: 'application/pdf',
+        buffer: largeBuffer
+      });
+
+      await continueButton.click();
+
+      await expect(page).toHaveURL(/\/manual-upload/);
+
+      errorSummary = page.locator('.govuk-error-summary');
+      await expect(errorSummary).toBeVisible();
+
+      errorLink = errorSummary.getByRole('link', { name: /file too large/i });
+      await expect(errorLink).toBeVisible();
+      await expect(errorLink).toHaveAttribute('href', '#file');
+
+      inlineErrorMessage = page.locator('#file').locator('..').locator('.govuk-error-message');
+      await expect(inlineErrorMessage).toBeVisible();
+      await expect(inlineErrorMessage).toContainText(/file too large, please upload file smaller than 2mb/i);
+    });
+
+    test('should show court name input with autocomplete initialized', async ({ page }) => {
+      await page.goto('/manual-upload');
+
+      const autocompleteInput = page.getByRole('combobox', { name: /court name or tribunal name/i });
+      await autocompleteInput.waitFor({ state: 'visible', timeout: 10000 });
+
+      await expect(autocompleteInput).toBeVisible();
+      await expect(autocompleteInput).toHaveAttribute('role', 'combobox');
+    });
+
+    test('should validate court name with empty, short, invalid inputs and preserve values', async ({ page }) => {
+      await page.goto('/manual-upload');
+
+      const courtInput = page.getByRole('combobox', { name: /court name or tribunal name/i });
+      await courtInput.waitFor({ state: 'visible', timeout: 10000 });
+
+      const fileInput = page.locator('input[name="file"]');
+      await fileInput.setInputFiles({
+        name: 'test.pdf',
+        mimeType: 'application/pdf',
+        buffer: Buffer.from('test content')
+      });
+
+      const continueButton = page.getByRole('button', { name: /continue/i });
+
+      await continueButton.click();
+
+      let errorSummary = page.locator('.govuk-error-summary');
+      let errorLink = errorSummary.getByRole('link', { name: /court name must be three characters or more/i });
+      await expect(errorLink).toBeVisible();
+
+      let inlineError = page.locator('.govuk-error-message').filter({ hasText: /court name must be three characters or more/i });
+      await expect(inlineError).toBeVisible();
+
+      await courtInput.fill('AB');
+      await continueButton.click();
+
+      errorSummary = page.locator('.govuk-error-summary');
+      await expect(errorSummary).toBeVisible();
+      errorLink = errorSummary.getByRole('link', { name: /court name must be three characters or more/i });
+      await expect(errorLink).toBeVisible();
+
+      inlineError = page.locator('#court-error.govuk-error-message');
+      await expect(inlineError).toBeVisible();
+      await expect(inlineError).toContainText(/court name must be three characters or more/i);
+
+      await courtInput.fill('Invalid Court Name That Does Not Exist');
+      await continueButton.click();
+
+      errorSummary = page.locator('.govuk-error-summary');
+      await expect(errorSummary).toBeVisible();
+      errorLink = errorSummary.getByRole('link', { name: /please enter and select a valid court/i });
+      await expect(errorLink).toBeVisible();
+
+      inlineError = page.locator('#court-error.govuk-error-message');
+      await expect(inlineError).toBeVisible();
+      await expect(inlineError).toContainText(/please enter and select a valid court/i);
+
+      const preservedCourtName = 'Invalid Court Name';
+      await courtInput.fill(preservedCourtName);
+      await continueButton.click();
+
+      await expect(courtInput).toHaveValue(preservedCourtName);
+    });
+
+    test('should validate date range and preserve all form data when validation fails', async ({ page }) => {
+      await page.goto('/manual-upload?locationId=1');
+      await page.waitForTimeout(1000);
+
+      const fileInput = page.locator('input[name="file"]');
+      await fileInput.setInputFiles({
+        name: 'test.pdf',
+        mimeType: 'application/pdf',
+        buffer: Buffer.from('test content')
+      });
+
+      await page.selectOption('select[name="listType"]', '1');
+      await page.fill('input[name="hearingStartDate-day"]', '15');
+      await page.fill('input[name="hearingStartDate-month"]', '06');
+      await page.fill('input[name="hearingStartDate-year"]', '2025');
+      await page.selectOption('select[name="sensitivity"]', 'PUBLIC');
+      await page.selectOption('select[name="language"]', 'ENGLISH');
+      await page.fill('input[name="displayFrom-day"]', '20');
+      await page.fill('input[name="displayFrom-month"]', '06');
+      await page.fill('input[name="displayFrom-year"]', '2025');
+      await page.fill('input[name="displayTo-day"]', '10');
+      await page.fill('input[name="displayTo-month"]', '06');
+      await page.fill('input[name="displayTo-year"]', '2025');
+
+      const continueButton = page.getByRole('button', { name: /continue/i });
+      await continueButton.click();
+
+      const errorSummary = page.locator('.govuk-error-summary');
+      await expect(errorSummary).toBeVisible();
+      const errorLink = errorSummary.getByRole('link', { name: /display to date must be after display from date/i });
+      await expect(errorLink).toBeVisible();
+
+      const inlineError = page.locator('#displayTo-error.govuk-error-message');
+      await expect(inlineError).toBeVisible();
+      await expect(inlineError).toContainText(/display to date must be after display from date/i);
+
+      await page.selectOption('select[name="sensitivity"]', 'PRIVATE');
+      await page.selectOption('select[name="language"]', 'WELSH');
+      await page.fill('input[name="displayFrom-day"]', '10');
+      await page.fill('input[name="displayTo-day"]', '20');
+
+      await continueButton.click();
+
+      await expect(page.locator('select[name="listType"]')).toHaveValue('1');
       await expect(page.locator('input[name="hearingStartDate-day"]')).toHaveValue('15');
       await expect(page.locator('input[name="hearingStartDate-month"]')).toHaveValue('06');
       await expect(page.locator('input[name="hearingStartDate-year"]')).toHaveValue('2025');
@@ -557,6 +509,270 @@ test.describe('Manual Upload Page', () => {
       await expect(page.locator('input[name="displayTo-day"]')).toHaveValue('20');
       await expect(page.locator('input[name="displayTo-month"]')).toHaveValue('06');
       await expect(page.locator('input[name="displayTo-year"]')).toHaveValue('2025');
+    });
+  });
+
+  test.describe('Manual Upload Summary Page', () => {
+    test('should display summary page with all elements and correct data', async ({ page }) => {
+      await navigateToSummaryPage(page);
+
+      await expect(page).toHaveURL(/\/manual-upload-summary\?uploadId=/);
+
+      const h1 = page.locator("h1");
+      await expect(h1).toHaveCount(1);
+      await expect(h1).toHaveText("File upload summary");
+
+      const mainHeading = page.getByRole("heading", { name: "File upload summary", level: 1 });
+      await expect(mainHeading).toBeVisible();
+
+      const subHeading = page.getByRole("heading", { name: "Check upload details", level: 2 });
+      await expect(subHeading).toBeVisible();
+      await expect(subHeading).toHaveText("Check upload details");
+
+      const backLink = page.locator(".govuk-back-link");
+      await expect(backLink).toBeVisible();
+
+      const summaryList = page.locator(".govuk-summary-list");
+      await expect(summaryList).toBeVisible();
+
+      const rows = page.locator(".govuk-summary-list__row");
+      await expect(rows).toHaveCount(7);
+
+      const keys = page.locator(".govuk-summary-list__key");
+      await expect(keys.nth(0)).toHaveText("Court name");
+      await expect(keys.nth(1)).toHaveText("File");
+      await expect(keys.nth(2)).toHaveText("List type");
+      await expect(keys.nth(3)).toHaveText("Hearing start date");
+      await expect(keys.nth(4)).toHaveText("Sensitivity");
+      await expect(keys.nth(5)).toHaveText("Language");
+      await expect(keys.nth(6)).toHaveText("Display file dates");
+
+      const values = page.locator(".govuk-summary-list__value");
+      await expect(values.nth(0)).not.toBeEmpty();
+      await expect(values.nth(1)).toContainText("test-document.pdf");
+      await expect(values.nth(2)).toContainText("Crown Daily List");
+      await expect(values.nth(3)).toContainText("23 October 2025");
+      await expect(values.nth(4)).toContainText("Public");
+      await expect(values.nth(5)).toContainText("English");
+      await expect(values.nth(6)).toContainText("20 October 2025 to 30 October 2025");
+
+      const changeLinks = page.locator(".govuk-summary-list__actions a");
+      await expect(changeLinks).toHaveCount(7);
+
+      for (let i = 0; i < 7; i++) {
+        await expect(changeLinks.nth(i)).toContainText("Change");
+      }
+
+      await expect(changeLinks.nth(0)).toHaveAttribute("href", "/manual-upload#court");
+      await expect(changeLinks.nth(1)).toHaveAttribute("href", "/manual-upload#file");
+      await expect(changeLinks.nth(2)).toHaveAttribute("href", "/manual-upload#listType");
+      await expect(changeLinks.nth(3)).toHaveAttribute("href", "/manual-upload#hearingStartDate-day");
+      await expect(changeLinks.nth(4)).toHaveAttribute("href", "/manual-upload#sensitivity");
+      await expect(changeLinks.nth(5)).toHaveAttribute("href", "/manual-upload#language");
+      await expect(changeLinks.nth(6)).toHaveAttribute("href", "/manual-upload#displayFrom-day");
+
+      const hiddenTexts = ["Court name", "File", "List type", "Hearing start date", "Sensitivity", "Language", "Display file dates"];
+      for (let i = 0; i < 7; i++) {
+        const visuallyHiddenText = changeLinks.nth(i).locator(".govuk-visually-hidden");
+        await expect(visuallyHiddenText).toHaveText(hiddenTexts[i]);
+      }
+
+      const confirmButton = page.getByRole("button", { name: "Confirm" });
+      await expect(confirmButton).toBeVisible();
+
+      const welshToggle = page.locator('a[href*="lng=cy"]');
+      await expect(welshToggle).not.toBeVisible();
+
+      const form = page.locator("form");
+      await expect(form).toBeVisible();
+      await expect(form).toHaveAttribute("method", "post");
+    });
+
+    test('should meet WCAG 2.2 AA standards on summary page', async ({ page }) => {
+      await navigateToSummaryPage(page);
+
+      const accessibilityScanResults = await new AxeBuilder({ page })
+        .withTags(["wcag2a", "wcag2aa", "wcag21a", "wcag21aa", "wcag22aa"])
+        .analyze();
+
+      expect(accessibilityScanResults.violations).toEqual([]);
+    });
+  });
+
+  test.describe('Manual Upload Success Page', () => {
+    test('should display success page with all elements and navigation links', async ({ page }) => {
+      await completeManualUploadFlow(page);
+
+      await expect(page).toHaveURL("/manual-upload-success");
+
+      const backLink = page.locator(".govuk-back-link");
+      await expect(backLink).not.toBeVisible();
+
+      const successPanel = page.locator(".govuk-panel");
+      await expect(successPanel).toBeVisible();
+
+      const title = page.locator(".govuk-panel__title");
+      await expect(title).toHaveText("File upload successful");
+
+      await expect(successPanel).toContainText("Your file has been uploaded");
+
+      const heading = page.getByRole("heading", { name: "What do you want to do next?" });
+      await expect(heading).toBeVisible();
+
+      const uploadLink = page.getByRole("link", { name: "Upload another file" });
+      await expect(uploadLink).toBeVisible();
+      await expect(uploadLink).toHaveAttribute("href", "/manual-upload");
+
+      const removeLink = page.getByRole("link", { name: "Remove file" });
+      await expect(removeLink).toBeVisible();
+      await expect(removeLink).toHaveAttribute("href", "/remove-list-search");
+
+      const homeLink = page.getByRole("link", { name: "Home" });
+      await expect(homeLink).toBeVisible();
+      await expect(homeLink).toHaveAttribute("href", "/admin-dashboard");
+
+      await uploadLink.click();
+      await expect(page).toHaveURL("/manual-upload");
+    });
+
+    test('should redirect to manual-upload if accessed directly without upload session', async ({ page }) => {
+      await page.goto("/manual-upload-success");
+      await expect(page).toHaveURL("/manual-upload");
+    });
+
+    test('should support Welsh language with correct translations', async ({ page }) => {
+      await completeManualUploadFlow(page);
+
+      const welshToggle = page.locator('a[href*="lng=cy"]');
+      await expect(welshToggle).not.toBeVisible();
+
+      await page.goto("/manual-upload-success?lng=cy");
+
+      const title = page.locator(".govuk-panel__title");
+      await expect(title).toHaveText("Wedi llwyddo i uwchlwytho ffeiliau");
+
+      const heading = page.getByRole("heading", { name: "Beth yr ydych eisiau ei wneud nesaf?" });
+      await expect(heading).toBeVisible();
+
+      const uploadLink = page.getByRole("link", { name: "uwchlwytho ffeil arall" });
+      await expect(uploadLink).toBeVisible();
+
+      const removeLink = page.getByRole("link", { name: "Dileu ffeil" });
+      await expect(removeLink).toBeVisible();
+
+      const homeLink = page.getByRole("link", { name: "Tudalen hafan" });
+      await expect(homeLink).toBeVisible();
+    });
+
+    test('should meet WCAG 2.2 AA standards on success page', async ({ page }) => {
+      await completeManualUploadFlow(page);
+
+      const accessibilityScanResults = await new AxeBuilder({ page })
+        .withTags(["wcag2a", "wcag2aa", "wcag21a", "wcag21aa", "wcag22aa"])
+        .analyze();
+
+      expect(accessibilityScanResults.violations).toEqual([]);
+    });
+
+
+    test('should not allow access after refreshing the page', async ({ page }) => {
+      await completeManualUploadFlow(page);
+
+      await page.reload();
+
+      await expect(page).toHaveURL("/manual-upload");
+    });
+
+    test('should allow multiple sequential uploads', async ({ page }) => {
+      await completeManualUploadFlow(page);
+      await expect(page).toHaveURL("/manual-upload-success");
+
+      await page.goto("/manual-upload?locationId=1");
+      await page.waitForTimeout(1000);
+
+      await page.selectOption('select[name="listType"]', "7");
+      await page.fill('input[name="hearingStartDate-day"]', "25");
+      await page.fill('input[name="hearingStartDate-month"]', "11");
+      await page.fill('input[name="hearingStartDate-year"]', "2025");
+      await page.selectOption('select[name="sensitivity"]', "PRIVATE");
+      await page.selectOption('select[name="language"]', "WELSH");
+      await page.fill('input[name="displayFrom-day"]', "24");
+      await page.fill('input[name="displayFrom-month"]', "11");
+      await page.fill('input[name="displayFrom-year"]', "2025");
+      await page.fill('input[name="displayTo-day"]', "26");
+      await page.fill('input[name="displayTo-month"]', "11");
+      await page.fill('input[name="displayTo-year"]', "2025");
+
+      const fileInput = page.locator('input[name="file"]');
+      await fileInput.setInputFiles({
+        name: "second-upload.pdf",
+        mimeType: "application/pdf",
+        buffer: Buffer.from("%PDF-1.4\nSecond upload content")
+      });
+
+      await page.getByRole("button", { name: /continue/i }).click();
+      await page.waitForURL(/\/manual-upload-summary\?uploadId=/);
+      await page.getByRole("button", { name: "Confirm" }).click();
+      await page.waitForURL("/manual-upload-success");
+
+      await expect(page).toHaveURL("/manual-upload-success");
+    });
+  });
+
+  test.describe('Responsive Design Across All Pages', () => {
+    test('should display correctly on mobile viewport throughout entire flow', async ({ page }) => {
+      await page.setViewportSize({ width: 375, height: 667 });
+
+      await page.goto('/manual-upload?locationId=1');
+      await page.waitForTimeout(1000);
+
+      const fileInput = page.locator('input[name="file"]');
+      await expect(fileInput).toBeVisible();
+      const courtInput = page.getByRole('combobox', { name: /court name or tribunal name/i });
+      await expect(courtInput).toBeVisible();
+
+      await page.selectOption('select[name="listType"]', "6");
+      await page.fill('input[name="hearingStartDate-day"]', "23");
+      await page.fill('input[name="hearingStartDate-month"]', "10");
+      await page.fill('input[name="hearingStartDate-year"]', "2025");
+      await page.selectOption('select[name="sensitivity"]', "PUBLIC");
+      await page.selectOption('select[name="language"]', "ENGLISH");
+      await page.fill('input[name="displayFrom-day"]', "20");
+      await page.fill('input[name="displayFrom-month"]', "10");
+      await page.fill('input[name="displayFrom-year"]', "2025");
+      await page.fill('input[name="displayTo-day"]', "30");
+      await page.fill('input[name="displayTo-month"]', "10");
+      await page.fill('input[name="displayTo-year"]', "2025");
+
+      await fileInput.setInputFiles({
+        name: "test-mobile.pdf",
+        mimeType: "application/pdf",
+        buffer: Buffer.from("%PDF-1.4\nTest mobile content")
+      });
+
+      await page.getByRole("button", { name: /continue/i }).click();
+      await page.waitForURL(/\/manual-upload-summary\?uploadId=/, { timeout: 10000 });
+
+      const h1Summary = page.locator("h1");
+      await expect(h1Summary).toBeVisible();
+
+      const summaryList = page.locator(".govuk-summary-list");
+      await expect(summaryList).toBeVisible();
+
+      const confirmButton = page.getByRole("button", { name: "Confirm" });
+      await expect(confirmButton).toBeVisible();
+
+      await confirmButton.click();
+      await page.waitForURL("/manual-upload-success", { timeout: 10000 });
+
+      const successPanel = page.locator(".govuk-panel");
+      await expect(successPanel).toBeVisible();
+
+      const heading = page.getByRole("heading", { name: "What do you want to do next?" });
+      await expect(heading).toBeVisible();
+
+      const links = page.locator(".govuk-list a");
+      await expect(links).toHaveCount(3);
     });
   });
 });
